Fix path assertions and bundling lists for devrig-common
Boundary-forced follow-ups to the moves:

- isWindows() relocated from BinLauncher.kt (where it was internal) to a
  new OsPlatform.kt in devrig-common — DevrigUserLauncher's default args
  call it; BinLauncher/AutoUpdater keep resolving it same-package.
- The migrateLegacyArchives test split out of HomePathsTest into a new
  npx-kt MigrateLegacyArchivesTest — its subject lives in
  ManagedBackend.kt, which stays in the devrig CLI.
- Backtick-quote KDoc references to CLI-side symbols (ensureBinLauncher,
  ENV_BIN_NO_AUTO_REGISTER) that no longer resolve across the module
  boundary.
- Add lib/devrig-common-<version>.jar to the verifyBundledLibraries
  expected lists of the devrig distZip and the plugin ZIP.
- Add devrig-common/src/main/kotlin to NoForbiddenPluginImportsTest's
  PRODUCTION_SOURCE_ROOTS (the module ships in the plugin ZIP).
- Add devrig-common to the root pluginCoreSubprojects guard set so
  ciBuildPluginTests must cover it.
