+++ devrig-common/src/main/kotlin/com/jonnyzzz/mcpSteroid/devrig/OsPlatform.kt
@@ -1,0 +1,5 @@
+/* Copyright 2025-2026 Eugene Petrenko (mcp@jonnyzzz.com); Copyright 2025-2026 JetBrains. Use of this source code is governed by the Apache 2.0 license. */
+package com.jonnyzzz.mcpSteroid.devrig
+
+fun isWindows(): Boolean =
+    System.getProperty("os.name").lowercase().contains("windows")
